Habits logic: -Took out data mock to a separate file -Created a Habit row to represent a row in the Habits table -State updates according to ui changes
